+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added a Missing Raid Buffs overlay filter
+- Added a Mine Only option for custom overlays
 - Added a configurable threat lead widget with role-aware status colors for target and focus frame styles
 
 ## [2.2.1](https://github.com/nivthefox/wow-nivui/releases/tag/v2.2.1) - 2026-08-25